refactor: Improve component structure and mobile responsiveness
- Rename Bento components to simpler names (AboutBento -> About, etc.)
- Consolidate education data into about data file
- Add mobile accordion view for Skills component with expand/collapse functionality
- Fix centering and scroll behavior for About and Skills components
- Ensure proper mobile/desktop responsive layouts with overflow handling
- Remove separate education module and integrate into about feature
