+++ src/features/about/index.ts
@@ -1,2 +1,2 @@
-export { default as AboutBento } from './components/AboutBento';
+export { default as AboutBento } from './components/About.tsx';
 export * from './data/about';
