--- src/features/education/index.ts
@@ -1,1 +1,0 @@
-export * from './data/education';
