+++ src/features/experience/index.ts
@@ -1,2 +1,2 @@
-export { default as ExperienceBento } from './components/ExperienceBento';
+export { default as ExperienceBento } from './components/Experience.tsx';
 export * from './data/experience';
