+++ src/features/projects/index.ts
@@ -1,2 +1,2 @@
-export { default as ProjectsBento } from './components/ProjectsBento';
+export { default as ProjectsBento } from './components/Projects.tsx';
 export * from './data/projects';
