+++ src/features/skills/index.ts
@@ -1,2 +1,2 @@
-export { default as SkillsBento } from './components/SkillsBento';
+export { default as SkillsBento } from './components/Skills.tsx';
 export * from './data/skills';
